Fix font size arguments.

+++ Inc/Font.h
@@ -29,8 +29,8 @@ public:
 		{ return fontFace; }
 
 	// Create an object with size information.
-	CFontSizeOwner CreateSizeObject( float pxSize ) const;
-	CFontSizeOwner CreateSizeObject( CVector2<float> pxSize ) const;
+	CFontSizeOwner CreateSizeObject( int pxSize ) const;
+	CFontSizeOwner CreateSizeObject( CVector2<int> pxSize ) const;
 
 	// Get the glyph structure for the given character code in UTF32BE encoding.
 	CGlyph GetGlyph( int charCode, CFontSizeView fontSize ) const;
@@ -73,8 +73,8 @@ public:
 	void Unload();
 
 	// Create an object with size information.
-	CFontSizeOwner CreateSizeObject( float pxSize ) const;
-	CFontSizeOwner CreateSizeObject( CVector2<float> pxSize ) const;
+	CFontSizeOwner CreateSizeObject( int pxSize ) const;
+	CFontSizeOwner CreateSizeObject( CVector2<int> pxSize ) const;
 	// Get the glyph structure for the given character code in UTF32BE encoding.
 	CGlyph GetGlyph( int charCode, CFontSizeView fontSize ) const;
